requirements.txt
# FastAPI 및 서버
fastapi==0.115.4
uvicorn[standard]==0.32.0

# Firebase
firebase-admin==6.5.0

# 환경 설정
python-dotenv==1.0.1

# 데이터 검증
pydantic==2.8.2
pydantic-settings==2.6.1

# 추가 유틸리티
python-multipart==0.0.12
email-validator==2.2.0

# MQTT (Arduino 센서 연동용 - 나중에 필요)
paho-mqtt==2.1.0

# WebSocket (Temi 로봇 연동용 - 나중에 필요)
websockets==13.1

# 👇 추가: HTTP 요청
httpx==0.27.0

bentoml==1.4.30

# OpenAI RAG
openai==1.54.0
requests==2.31.0